OTHER: Moved PERMC_SPEC to utils.py
Moved the permc_spec constant to utils so that morphological.mpls can also use the value.
