+++ pybaselines/morphological.py
@@ -20,7 +20,7 @@ from scipy.ndimage import (grey_closing, grey_dilation, grey_erosion,
 from scipy.sparse.linalg import spsolve
 
 from ._algorithm_setup import _optimize_window, _setup_morphology, _setup_whittaker
-from .utils import pad_edges, padded_convolve, relative_difference
+from .utils import PERMC_SPEC, pad_edges, padded_convolve, relative_difference
 
 
 # make _optimize_window available to users and document it so that
@@ -213,7 +213,7 @@ def mpls(data, half_window=None, lam=1e6, p=0.0, diff_order=2, tol=1e-3, max_ite
             w[index] = 1 - p
 
     _, diff_matrix, weight_matrix, weight_array = _setup_whittaker(y, lam, diff_order, w)
-    baseline = spsolve(weight_matrix + diff_matrix, weight_array * y)
+    baseline = spsolve(weight_matrix + diff_matrix, weight_array * y, permc_spec=PERMC_SPEC)
 
     residual = y - baseline
     params = {
